backend/pyproject.toml
[project]
name = "cropvision-backend"
version = "0.1.0"
description = "农作物病害智能识别系统后端服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]